+++ src/main/java/BusinessLayer/Policies/DiscountPolicies/PolicyTypes/MinQuantityDiscountPolicy.java
@@ -24,7 +24,7 @@ public class MinQuantityDiscountPolicy extends BaseDiscountPolicy {
     private final boolean allowNone;
 
     public MinQuantityDiscountPolicy(int policyId, Product product, int minQuantity, boolean allowNone,int store_id,int discount_id) throws Exception {
-        super(policyId,store_id,discount_id);
+        super(policyId,store_id, (long) discount_id);
         if (minQuantity <= 0) {
             logger.error("Min quantity must be larger than 0 but is " + minQuantity);
             throw new Exception("Min quantity must be larger than 0");
